+++ app/src/main/java/com/brewthings/app/data/repository/RaptPillRepository.kt
@@ -56,4 +56,9 @@ class RaptPillRepository(
         dao.observeData(macAddress).map { data ->
             data.map { it.toModelItem() }
         }
+
+    fun observeOG(macAddress: String): Flow<RaptPillData?> =
+        dao.observeOG(macAddress).map {
+            it?.toModelItem()
+        }
 }
