+++ app/src/main/java/com/brewthings/app/di/Koin.kt
@@ -1,7 +1,6 @@
 package com.brewthings.app.di
 
 import com.brewthings.app.data.ble.RaptPillScanner
-import com.brewthings.app.data.repository.RaptPillInsightsRepository
 import com.brewthings.app.data.repository.RaptPillRepository
 import com.brewthings.app.data.storage.RaptPillDatabase
 import com.brewthings.app.ui.screens.graph.GraphScreenViewModel
@@ -15,7 +14,6 @@ val appModule = module {
     single { RaptPillDatabase.create(context = androidContext()) }
     factory { get<RaptPillDatabase>().raptPillDao() }
     factory { RaptPillRepository(scanner = get(), dao = get()) }
-    factory { RaptPillInsightsRepository(macAddress = get(), dao = get()) }
     viewModel { ScanningScreenViewModel() }
     viewModel { GraphScreenViewModel() }
 }
